Add File Loader implementation
This implementation is largely untested and, given the high amount of
interfaces and modularity, is prone to the presence of bugs.
Nonetheless, consumers may rely on it and report any problems they find.

Testing will be done directly inside Boson after the addition of the
Tags API or the registering of all types of recipes through JSON,
whichever comes first.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>