Fix loading issues and complete API
First off, let's avoid running the scan more than once. Also, every
ModContainer will now be considered for paths, even if a valid path does
not exist. This is done so that Visitors have an accurate representation
of the places that the locator visits and checks for files.

Also, changed the ChainingProgressVisitor implementation so that it
supports multiple ProgressVisitors without having to nest them in weird
ways.

The phases API was also uniformed to the configuration API, with the
name of the phase now being required and part of a method invocation to
construct the phase itself.

Locators can now clean up after their resources at the end of the
loading process. Since locators are initialized only once per loader
(they're not Phase-specific), they may need to clean up resources and
they had no way to do that before apart from guessing or requiring a
special-cased visitors. The clean up method is now invoked right before
the loading completion.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>