Add data-pack-like locator to prefab
What this locator does is loop over all mod containers looking for all
subdirectories of the kind that we're looking for and parsing those.
This is particularly useful if you need data-pack-like support, i.e.
allow overriding or adding of elements or entries.

This is highly discouraged for registry entries or anything that is
prone to breaking

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>