Move from Item to ItemStack and from Block to BlockState
Also add some conversion functions from those types

This allows the tag system to properly support tags with metadata and
with variants, that would otherwise be completely unsupported with the
old system (i.e. raw Item and Block).

Raw types are in fact perfect in a flattened environment, i.e. where
metadata or blockstate variant does not carry information, but currently
this is not the case

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>