Remove exposure of potentially incorrect tag types in Prefab
Instead of creating the tag types in the prefab section of the mod, we
do that in the implementation only. Instead, the exposed bits of the
prefab are just wrappers around calls to the tag type registry, which is
nevertheless fast. The main difference is that it makes for
better-looking code.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>