Move tag-type registration event to pre-init
CT registers the recipes on the registry event for IRecipes, which is
correct on their part. Unfortunately this comes too early for us, so
moving tag type registration to pre-init is the most sensible solution.
We don't want to register tags on pre-init though, since blocks and
items don't exist yet. Also, we don't want to have them exist as a forge
registry since that doesn't make sense for the way they're actually
loaded (and due to the fact that tags can reference any other kind of
forge registry, so it would be an ordering issue again).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>