Add an equality check to Tag Types
This allows TagType creators to specify a different equality behavior
rather than relying on `==` (i.e. `equals`), for special cases, like
with `ItemStack`s not overriding equals but rather relying on identity.

Also, changed the implementation to account for these changes and to use
the specified equality check in the implementation to avoid duplicate
items getting added to the `elements` set. Mutable sets are left
untouched due to performance and code concerns.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>